+++ android/src/main/java/com/web3authreactnativesdk/Web3authReactNativeSdkPackage.kt
@@ -1,4 +1,4 @@
-package com.openloginreactnativesdk
+package com.web3authreactnativesdk
 
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
@@ -6,9 +6,9 @@ import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 
 
-class OpenloginReactNativeSdkPackage : ReactPackage {
+class Web3authReactNativeSdkPackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-        return listOf(OpenloginReactNativeSdkModule(reactContext))
+        return listOf(Web3authReactNativeSdkModule(reactContext))
     }
 
     override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
